Put trees everywhere, fix lighting.
The height from the heightmap is wrong, but I need to leave right now, so...
checkpointing.

+++ src/base/Terrain.h
@@ -3,7 +3,7 @@
 #include "base/Program.h"
 #include "geometry/Mesh.h"
 
-const float TERRAIN_DIMENSIONS = 100.0f;
+const size_t TERRAIN_DIMENSIONS = 100;
 
 class Terrain final : public Mesh {
   Terrain(std::vector<Vertex>&& vertices,
